Extract metrics into their own files following the example of traces
