+++ src/__assets__/output_success.txt
@@ -355,8 +355,8 @@
         'pull_request.number': 9011,
         'workflow.event': 'pull_request_target'
       },
-      startTime: [ 1770734182, 581000000 ],
-      endTime: [ 1770734182, 584000000 ],
+      startTime: [ 1770739184, 360000000 ],
+      endTime: [ 1770739184, 361000000 ],
       value: 563000
     }
   ]
@@ -378,8 +378,8 @@
         'pull_request.number': 9011,
         'workflow.event': 'pull_request_target'
       },
-      startTime: [ 1770734182, 581000000 ],
-      endTime: [ 1770734182, 585000000 ],
+      startTime: [ 1770739184, 360000000 ],
+      endTime: [ 1770739184, 362000000 ],
       value: 563000
     }
   ]
